03.02.23: Day3

+++ src/main/kotlin/Attack.kt
@@ -1,1 +1,1 @@
-class Attack (val name: String, val damage: Double, val type: String)
+class Attack (val name: String, val damage: Int, val type: String)
